--- src/components/AppNavBar/AppNavBar.module.css
@@ -1,7 +1,0 @@
-.nav {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 10px 20px;
-  background-color: #a1a1a1;
-}
