+++ src/components/Logo.jsx
@@ -1,0 +1,12 @@
+import { Link } from "react-router-dom";
+import styles from "./Logo.module.css";
+
+function Logo() {
+  return (
+    <Link to="/">
+      <img src="/logo.png" alt="Traveler logo" className={styles.logo} />
+    </Link>
+  );
+}
+
+export default Logo;
