+++ src/components/Logo.module.css
@@ -1,0 +1,8 @@
+.logo {
+  height: 5.2rem;
+  margin: 1rem;
+  padding: 0.5rem;
+  margin-left: 8rem;
+  transform: scale(1.7);
+  margin-top: 5rem;
+}
