+++ src/components/PagesNavBar.jsx
@@ -1,18 +1,22 @@
 import { NavLink } from "react-router-dom";
 import styles from "./PagesNavBar.module.css";
+import Logo from "./Logo";
 
 function PagesNavBar() {
   return (
     <nav className={styles.nav}>
+      <Logo />
       <ul>
-        <li>
-          <NavLink to="/">Home</NavLink>
-        </li>
         <li>
           <NavLink to="/pricing">Pricing</NavLink>
         </li>
         <li>
           <NavLink to="/product">Product</NavLink>
+        </li>
+        <li>
+          <NavLink to="/login" className={styles.ctaLink}>
+            Login
+          </NavLink>
         </li>
       </ul>
     </nav>
