+++ src/components/PagesNavBar.module.css
@@ -2,6 +2,7 @@
   display: flex;
   align-items: center;
   justify-content: space-between;
+  background-color: var(--color-bg--main);
 }
 
 .nav ul {
@@ -28,7 +29,7 @@
 a.ctaLink:link,
 a.ctaLink:visited {
   background-color: var(--color-brand--2);
-  color: var(--color-dark--0);
+  color: var(--color-bg-secondary--1);
   padding: 0.8rem 2rem;
   border-radius: 7px;
 }
