--- src/components/PagesNavBar/PagesNavBar.module.css
@@ -1,18 +1,0 @@
-.nav {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 10px 20px;
-  background-color: #c3c3c3;
-}
-
-.nav ul {
-  list-style: none;
-  display: flex;
-  gap: 20px;
-}
-
-.nav :global(.active) {
-  font-weight: bold;
-  background-color: #989898;
-}
