+++ src/pages/Home.jsx
@@ -1,12 +1,24 @@
 import { Link } from "react-router-dom";
+import styles from "./Home.module.css";
 
-function Home() {
+export default function Home() {
   return (
-    <div>
-      <h1>Traveler</h1>
-      <Link to="/app">To the app</Link>
-    </div>
+    <main className={styles.homepage}>
+      <section>
+        <h1>
+          You travel the world.
+          <br />
+          WorldWise keeps track of your adventures.
+        </h1>
+        <h2>
+          A world map that tracks your footsteps into every city you can think
+          of. Never forget your wonderful experiences, and show your friends how
+          you have wandered the world.
+        </h2>
+        <Link to="/app" className="cta">
+          Start your adventure
+        </Link>
+      </section>
+    </main>
   );
 }
-
-export default Home;
