+++ src/pages/Login.module.css
@@ -1,12 +1,12 @@
 .login {
   margin: 2.5rem;
   padding: 2.5rem 5rem;
-  background-color: var(--color-dark--1);
+  background-color: var(--color-bg--main);
   min-height: calc(100vh - 5rem);
 }
 
 .form {
-  background-color: var(--color-dark--2);
+  background-color: var(--color-bg-secondary--0);
   border-radius: 7px;
   padding: 2rem 3rem;
   width: 100%;
