+++ src/pages/PageNotFound.jsx
@@ -1,5 +1,7 @@
-function PageNotFound() {
-  return <div>Not Found! :(</div>;
+export default function PageNotFound() {
+  return (
+    <div>
+      <h1>Page not found 😢</h1>
+    </div>
+  );
 }
-
-export default PageNotFound;
