+++ src/pages/Product.module.css
@@ -1,7 +1,7 @@
 .product {
   margin: 2.5rem;
   padding: 2.5rem 5rem;
-  background-color: var(--color-dark--1);
+  background-color: var(--color-bg--main);
   min-height: calc(100vh - 5rem);
 }
