+++ starter/components/Button.module.css
@@ -12,7 +12,7 @@
 .primary {
   font-weight: 700;
   background-color: var(--color-brand--2);
-  color: var(--color-dark--1);
+  color: var(--color-bg--main);
 }
 
 .back {
@@ -30,6 +30,6 @@
   left: 50%;
   transform: translateX(-50%);
   background-color: var(--color-brand--2);
-  color: var(--color-dark--1);
+  color: var(--color-bg--main);
   box-shadow: 0 0.4rem 1.2rem rgba(36, 42, 46, 0.16);
 }
