+++ starter/components/CityItem.module.css
@@ -5,7 +5,7 @@
   gap: 1.6rem;
   align-items: center;
 
-  background-color: var(--color-dark--2);
+  background-color: var(--color-bg-secondary--0);
   border-radius: 7px;
   padding: 1rem 2rem;
   border-left: 5px solid var(--color-brand--2);
@@ -40,7 +40,7 @@
   aspect-ratio: 1;
   border-radius: 50%;
   border: none;
-  background-color: var(--color-dark--1);
+  background-color: var(--color-bg--main);
   color: var(--color-light--2);
   font-size: 1.6rem;
   font-weight: 400;
@@ -50,5 +50,5 @@
 
 .deleteBtn:hover {
   background-color: var(--color-brand--1);
-  color: var(--color-dark--1);
+  color: var(--color-bg--main);
 }
