+++ starter/components/CountryItem.module.css
@@ -7,7 +7,7 @@
   font-size: 1.7rem;
   font-weight: 600;
 
-  background-color: var(--color-dark--2);
+  background-color: var(--color-bg-secondary--0);
   border-radius: 7px;
   padding: 1rem 2rem;
   border-left: 5px solid var(--color-brand--1);
