--- starter/components/Logo.module.css
@@ -1,3 +1,0 @@
-.logo {
-  height: 5.2rem;
-}
