+++ starter/components/Sidebar.module.css
@@ -1,6 +1,6 @@
 .sidebar {
   flex-basis: 56rem;
-  background-color: var(--color-dark--1);
+  background-color: var(--color-bg--main);
   padding: 3rem 5rem 3.5rem 5rem;
 
   display: flex;
