+++ starter/components/SpinnerFullPage.module.css
@@ -1,5 +1,5 @@
 .spinnerFullpage {
   margin: 2.5rem;
   height: calc(100vh - 5rem);
-  background-color: var(--color-dark--1);
+  background-color: var(--color-bg--main);
 }
