--- starter/pages/PageNotFound.jsx
@@ -1,7 +1,0 @@
-export default function PageNotFound() {
-  return (
-    <div>
-      <h1>Page not found 😢</h1>
-    </div>
-  );
-}
